Move cliqueSealBlock functionality into standalone constructor function
