+++ packages/block/src/constructors.ts
@@ -7,6 +7,7 @@ import {
   createTxFromTxData,
 } from '@ethereumjs/tx'
 import {
+  BIGINT_27,
   CLRequestFactory,
   ConsolidationRequest,
   DepositRequest,
@@ -16,6 +17,8 @@ import {
   bigIntToHex,
   bytesToHex,
   bytesToUtf8,
+  concatBytes,
+  ecsign,
   equalsBytes,
   fetchFromProvider,
   getProvider,
@@ -24,6 +27,7 @@ import {
   isHexString,
 } from '@ethereumjs/util'
 
+import { _requireClique } from './consensus/clique.js'
 import { createBlockFromRpc } from './from-rpc.js'
 import {
   genRequestsTrieRoot,
@@ -32,7 +36,14 @@ import {
   valuesArrayToHeaderData,
 } from './helpers.js'
 
-import { Block, BlockHeader, executionPayloadFromBeaconPayload } from './index.js'
+import {
+  Block,
+  BlockHeader,
+  CLIQUE_EXTRA_SEAL,
+  CLIQUE_EXTRA_VANITY,
+  cliqueSigHash,
+  executionPayloadFromBeaconPayload,
+} from './index.js'
 
 import type { BeaconPayloadJson } from './from-beacon-payload.js'
 import type {
@@ -514,4 +525,32 @@ export async function createBlockFromBeaconPayloadJson(
 ): Promise<Block> {
   const executionPayload = executionPayloadFromBeaconPayload(payload)
   return createBlockFromExecutionPayload(executionPayload, opts)
+}
+
+/**
+ *  Method to retrieve a sealed block from the header and clique signer key
+ * @param header block header to be used for creating sealed block
+ * @param cliqueSigner clique signer key to be used for creating sealed block
+ * @returns the block constructed block
+ */
+export function createSealedCliqueBlock(header: BlockHeader, cliqueSigner: Uint8Array): Uint8Array {
+  // Ensure extraData is at least length CLIQUE_EXTRA_VANITY + CLIQUE_EXTRA_SEAL
+  const minExtraDataLength = CLIQUE_EXTRA_VANITY + CLIQUE_EXTRA_SEAL
+  if (header.extraData.length < minExtraDataLength) {
+    const remainingLength = minExtraDataLength - header.extraData.length
+    ;(header.extraData as any) = concatBytes(header.extraData, new Uint8Array(remainingLength))
+  }
+
+  _requireClique(header, 'cliqueSealBlock')
+
+  const ecSignFunction = header.common.customCrypto?.ecsign ?? ecsign
+  const signature = ecSignFunction(cliqueSigHash(header), cliqueSigner)
+  const signatureB = concatBytes(signature.r, signature.s, bigIntToBytes(signature.v - BIGINT_27))
+
+  const extraDataWithoutSeal = header.extraData.subarray(
+    0,
+    header.extraData.length - CLIQUE_EXTRA_SEAL,
+  )
+  const extraData = concatBytes(extraDataWithoutSeal, signatureB)
+  return extraData
 }
